fix(ci): runIfScript moves to @mailwoman/core/utils — core scripts can't import from mailwoman (graph inversion)
The operator's runIfScript standardization pointed two core/data scripts at mailwoman/sdk/scripting — core builds before mailwoman on a clean tree, so CI broke while stale local out/ masked it (the recurring class). The implementation now lives in core/utils/scripting (es-main joins core deps); mailwoman/sdk/scripting stays as the stable path via named re-exports; the core scripts self-reference @mailwoman/core/utils (precedented, clean-tree-verified, keeps the documented source-run contract working).
