+++ core/package.json
@@ -69,6 +69,7 @@
 		"change-case": "^5.4.4",
 		"d3-color": "^3.1.0",
 		"d3-scale-chromatic": "^3.1.0",
+		"es-main": "^1.4.0",
 		"fast-glob": "^3.3.3",
 		"http-status-codes": "^2.3.0",
 		"json-colorizer": "^3.0.1",
