+++ core/utils/index.ts
@@ -9,3 +9,4 @@ export * from "./repo.js"
 export * from "./sealed-db.js"
 export * from "./python-json.js"
 export * from "./python-random.js"
+export * from "./scripting.js"
